Added new method to the ParkingRateDao to support Android app's functionality. Also changed all Dao object's id from int to long, along with DB schema and script changes.

+++ ParqDBdesign/src/main/resources/com/parq/db/design/DB_Creation_Script.sql
@@ -5,7 +5,7 @@ CREATE DATABASE parq;
 USE parq;
 
 CREATE TABLE user
-(user_id INT NOT NULL AUTO_INCREMENT,  
+(user_id BIGINT NOT NULL AUTO_INCREMENT,  
  password TEXT(64) NOT NULL, 
  email TEXT(64) NOT NULL,
  phone_number TEXT(64),
@@ -14,8 +14,8 @@ CREATE TABLE user
  PRIMARY KEY (user_id));
  
 CREATE TABLE paymentaccount
-(account_id INT NOT NULL AUTO_INCREMENT,  
- user_id INT NOT NULL,
+(account_id BIGINT NOT NULL AUTO_INCREMENT,  
+ user_id BIGINT NOT NULL,
  customer_id TEXT(64), 
  payment_method_id TEXT(64),
  cc_stub TEXT(64),
@@ -26,7 +26,7 @@ CREATE TABLE paymentaccount
  FOREIGN KEY (user_id) references user(user_id));
  
 CREATE TABLE userrole
-(userrole_id INT NOT NULL AUTO_INCREMENT,
+(userrole_id BIGINT NOT NULL AUTO_INCREMENT,
  role_name TEXT(64) NOT NULL,
  role_desc TEXT(256),
  parking_rate DECIMAL(5,3),
@@ -34,7 +34,7 @@ CREATE TABLE userrole
  PRIMARY KEY (userrole_id));
 
 CREATE TABLE admin
-(admin_id INT NOT NULL AUTO_INCREMENT,
+(admin_id BIGINT NOT NULL AUTO_INCREMENT,
  username TEXT(64) NOT NULL,
  password TEXT(64) NOT NULL,
  email TEXT(64) NOT NULL,
@@ -43,7 +43,7 @@ CREATE TABLE admin
  PRIMARY KEY (admin_id));
  
 CREATE TABLE client
-(client_id INT NOT NULL AUTO_INCREMENT,
+(client_id BIGINT NOT NULL AUTO_INCREMENT,
  name TEXT(64) NOT NULL,
  address TEXT(256),
  client_desc TEXT(256),
@@ -52,17 +52,17 @@ CREATE TABLE client
  PRIMARY KEY (client_id)); 
  
 CREATE TABLE adminrole
-(adminrole_id INT NOT NULL AUTO_INCREMENT,
+(adminrole_id BIGINT NOT NULL AUTO_INCREMENT,
  role_name TEXT(64) NOT NULL,
  role_desc TEXT(256),
  lastupdatedatetime TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP, 
  PRIMARY KEY (adminrole_id)); 
  
 CREATE TABLE adminclientrelationship
-(ac_rel_id INT NOT NULL AUTO_INCREMENT,
- admin_id INT NOT NULL,
- client_id INT NOT NULL,
- adminrole_id INT NOT NULL,
+(ac_rel_id BIGINT NOT NULL AUTO_INCREMENT,
+ admin_id BIGINT NOT NULL,
+ client_id BIGINT NOT NULL,
+ adminrole_id BIGINT NOT NULL,
  lastupdatedatetime TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP, 
  PRIMARY KEY (ac_rel_id),
  FOREIGN KEY (admin_id) references admin(admin_id),
@@ -70,9 +70,9 @@ CREATE TABLE adminclientrelationship
  FOREIGN KEY (adminrole_id) references adminrole(adminrole_id));
  
 CREATE TABLE parkinglocation
-(location_id INT NOT NULL AUTO_INCREMENT,
+(location_id BIGINT NOT NULL AUTO_INCREMENT,
  location_identifier TEXT(16) NOT NULL,
- client_id INT NOT NULL,
+ client_id BIGINT NOT NULL,
  location_name TEXT(64),
  is_deleted BOOLEAN DEFAULT FALSE,
  lastupdatedatetime TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP, 
@@ -80,8 +80,8 @@ CREATE TABLE parkinglocation
  FOREIGN KEY (client_id) references client(client_id));
  
 CREATE TABLE geolocation
-(geolocation_id INT NOT NULL AUTO_INCREMENT,
- location_id INT NOT NULL,
+(geolocation_id BIGINT NOT NULL AUTO_INCREMENT,
+ location_id BIGINT NOT NULL,
  latitude DOUBLE NOT NULL,
  longitude DOUBLE NOT NULL,
  lastupdatedatetime TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP, 
@@ -89,9 +89,9 @@ CREATE TABLE geolocation
  FOREIGN KEY (location_id) references parkinglocation(location_id));
  
 CREATE TABLE parkingspace
-(space_id INT NOT NULL AUTO_INCREMENT,
+(space_id BIGINT NOT NULL AUTO_INCREMENT,
  space_identifier TEXT(16) NOT NULL,
- location_id INT NOT NULL,
+ location_id BIGINT NOT NULL,
  space_name TEXT(64),
  parking_level TEXT(16),
  is_deleted BOOLEAN DEFAULT FALSE,
@@ -100,27 +100,27 @@ CREATE TABLE parkingspace
  FOREIGN KEY (location_id) references parkinglocation(location_id));
 
 CREATE TABLE parkingrate
-(rate_id INT NOT NULL AUTO_INCREMENT,
- location_id INT NOT NULL,
+(rate_id BIGINT NOT NULL AUTO_INCREMENT,
+ location_id BIGINT NOT NULL,
  parking_rate_cents INT NOT NULL,
  time_increment_mins INT NOT NULL,
  priority INT NOT NULL,
  min_park_mins INT,
  max_park_mins INT,
- space_id INT,
+ space_id BIGINT,
  lastupdatedatetime TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP, 
  PRIMARY KEY (rate_id),
  FOREIGN KEY (location_id) references parkinglocation(location_id),
  FOREIGN KEY (space_id) references parkingspace(space_id));
  
 CREATE TABLE userclientrelationship
-(uc_rel_id INT NOT NULL AUTO_INCREMENT,
- user_id INT NOT NULL,
- client_id INT NOT NULL,
- userrole_id INT NOT NULL,
+(uc_rel_id BIGINT NOT NULL AUTO_INCREMENT,
+ user_id BIGINT NOT NULL,
+ client_id BIGINT NOT NULL,
+ userrole_id BIGINT NOT NULL,
  priority INT NOT NULL,
- location_id INT,
- space_id INT,
+ location_id BIGINT,
+ space_id BIGINT,
  lastupdatedatetime TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP, 
  PRIMARY KEY (uc_rel_id),
  FOREIGN KEY (user_id) references user(user_id),
@@ -130,9 +130,9 @@ CREATE TABLE userclientrelationship
  FOREIGN KEY (space_id) references parkingspace(space_id));
 
 CREATE TABLE parkinginstance
-(parkinginst_id INT NOT NULL AUTO_INCREMENT,
- user_id INT NOT NULL,
- space_id INT NOT NULL,
+(parkinginst_id BIGINT NOT NULL AUTO_INCREMENT,
+ user_id BIGINT NOT NULL,
+ space_id BIGINT NOT NULL,
  parkingrefnumber TEXT(64) NOT NULL,
  park_began_time DATETIME NOT NULL,
  park_end_time DATETIME NOT NULL,
@@ -143,10 +143,10 @@ CREATE TABLE parkinginstance
  FOREIGN KEY (space_id) references parkingspace(space_id));
  
 CREATE TABLE payment
-(payment_id INT NOT NULL AUTO_INCREMENT,
- parkinginst_id INT NOT NULL,
+(payment_id BIGINT NOT NULL AUTO_INCREMENT,
+ parkinginst_id BIGINT NOT NULL,
  payment_type TEXT(64) NOT NULL,
- account_id INT,
+ account_id BIGINT,
  payment_ref_num TEXT(64) NOT NULL,
  payment_datetime DATETIME NOT NULL,
  amount_paid_cents INT NOT NULL,
